initial threadmanager method split

+++ FileDeleter.java
@@ -9,10 +9,10 @@ public class FileDeleter {
     String filePath;
     String baseFileName;
     String extension;
-    int numberOfFiles;
+    long numberOfFiles;
     File toDelete;
     
-    public FileDeleter(String filePath, String baseFileName, String extension, int numberOfFiles){
+    public FileDeleter(String filePath, String baseFileName, String extension, long numberOfFiles){
         this.filePath = filePath;
         this.baseFileName = baseFileName;
         this.extension = extension;
@@ -22,7 +22,7 @@ public class FileDeleter {
     // deletes files specified in constructor
     public void delete(){
         String path;
-        for (int i = 1; i < numberOfFiles; i++) {
+        for (long i = 1; i < numberOfFiles; i++) {
             path = filePath + baseFileName + i + extension;
             toDelete = new File(path);
             toDelete.delete();
